[TK] Add a TestLaunchContext for generating test dispatch IRs (#429)
This patch changes the way we do grid bindings. It is now expected that
the IndexingContext will have the workload values and the grid on
construction will try to build it's dims using them. Currently, all
symbols are assumed to have a constant value, so the grid is able to map
workgroup calculations to constants.

Ideally, we would have support to parameterize the kernel by some
symbols, which would be passed as workload bindings (We have no support
for this currently).