[TK] Add support to generate tk operations for torch.export (#504)
* Adds `tkl.gen.kernel` decorator to generate a registered/callable
kernel.
* The kernel can be used in both AOT and eager contexts.
* Adds some support to kernel registration for uniqueing names like
these that are not user-facing and can be defined multiple times.
* Changes the CompiledModule phasing to include a custom op expansion
phase. Also defaults it to `IMPORT` instead of `FULL` in a couple of low
level places where it still had the old default. This is now consistent
with expected, canonical usage but still allows importing to IREE's
internal IR if desired (i.e. `FULL`).

---------

Co-authored-by: Stella Laurenzo <stellaraccident@gmail.com>