[export] Add a helper that lets us export multiple entrypoints from an nn.Module (#511)
I really hope that the `torch.export` folks consider this a real use
case to solve for, but for now it seems like it is us and the executorch
folks doing gross hacks like this to support export of multiple
functions from a module stack.

I went the extra mile on this a bit and went ahead to implement
save/load on these stacks, which gives us the ability to do FX tracing
once and then iterate on the result to compile. It should be an easier
workflow (for dev) considering that Dynamo is so slow for real models.

This is all bleeding edge stuff on Torch, but I need it for LLM so will
deal with the sharp edge.