Upgrade to PyTorch 2.3. (#546)
As discussed on Discord, this is a significant upgrade because it is the
first stable release that has a fully functional `torch.export.export`
with the preferred dynamic shapes support. It is also just prior to
nightlies that completely remove support for the old constraints based
API, so is therefore a good point to stop for a moment and support both
styles.

This patch makes a number of API changes:

* Issues deprecation warnings if the `constraints=` keyword for jittable
is used, otherwise not passing it to PyTorch. This should make jittable
not immediately incompatible with later nightlies unless if that feature
is used.
* Adds the ability for a `CompiledModule` to directly have an attribute
of a `torch.export.ExportedProgram`, allowing the user to pre-export
with Torch and then construct a compiled module from that (vs the
`jittable` approach where the `CompiledModule` API was directly invoking
Torch internals to do so). This defaults to exporting as `public` if
given a name not starting with an underscore and private otherwise.
Private ExportedPrograms can be called from procedures just as with
`jittable`.
* `shark_turbine.aot.export()` now accepts either an `CompiledModule`,
`nn.Module`, a or a `torch.export.ExportedProgram`. For the last two, a
new `external_params=` bool is available to control whether parameters
are inlined or externalized. For an `nn.Module` arguments corresponding
to `torch.export.export` are added. Internally, for an `nn.Module`, it
simply calls `torch.export.export`. `jittable` is no longer used
internally.

Some attempt has been made to be backwards compatible with Torch 2.1.0.
New features will not work, but we should be able to support a short
buffer window where older pinned systems are not completely broken. The
repository prior to this patch will be branched to `torch_2.1`.

Breaking changes:

* ops.iree.trace_tensors (plural) had to be removed because the PyTorch
auto functionalization thing has a TODO around lists of tensors. We can
add a wrapper that takes a list and invokves trace_tensors multiple
times and/or ass a `functional_trace_tensors` which works a bit better
with the infra.
* stateless_llama_test.py::test_rerotated_torch_comparison marked as
expectedFailure. Filed #560