Add device bridge support for HIP and CUDA. (#3)
* Adds a custom IREE "_test_add" op which exercises actual code
generation.
* Reworks the device management layer to:
* Initialize our `Device` wrapper to maintain the correspondence between
IREE HalDevice and PyTorch device.
  * Annotated Device with dlpack_device_type_code.
* Detects whether the Torch device is for real-CUDA or AMDGPU/HIP CUDA
and interfaces to the outside world correctly based on this.
* Auto-detects the AMDGPU chip and the CUDA SM version and arranges for
the JIT compiler to use that.
* Dynamically enables custom kernel registration for CUDA if it is
available.

Limitations (for now):

* IREE's dlpack interop only supports contiguous tensors. Lifting this
requires further interfacing with the compiler to specialize on
different strided layouts.
* Device synchronization is using dlpack's implicit default-stream
synchronization. Since we are currently still JIT compiling in
synchronous mode, we are skating by on this. Some additional hooks and
APIs are needed to properly place stream events to do it for real.
* ROCM builds of torch seem to not have an easy way to map a device back
to its UUID, and I didn't have a CUDA device handy to test the ways this
is done on CUDA, so both are just relying on enumeration order on multi
device systems. I kept the correspondence to a single place in the code
and think I know how to fix this, but it will require some poking.


Depends on an IREE bump that includes:
https://github.com/iree-org/iree/pull/17131 (but can go in without it
since that only unblocks HIP/CUDA, which were not supported yet anyway)