Add initial CI (#4)
This commit brings over all the CI from SHARK-Turbine and configures it
properly for iree-turbine. I also add parallelization to the model
testing, so that the cpu model tests run on ubuntu 64 core and gpu tests
run on mi250. This way we only hog the mi250 for 4 mins compared to the
~30 mins in SHARK Turbine (we use mi250 for gpu model regression testing
on iree PRs so wanted to reduce time being used to only necessity). I
didn't see anywhere in iree-org that solely depends on ubuntu 64 core so
I don't think we would be blocking anywhere by hogging for while. But,
let me know if we want further parallelization due to github runner cost
of using 64 core (I can try and see if we can get some of the sd tests
running on 32 or 16 core even)