CI Fixes for test_models and SHARK testing (#31)
This commit makes the necessary changes to both the workflows to get the
CI working properly. The SHARK workflow file is temporarily using a
branch that I created in nod-ai/SHARK to resolve some versioning issues.
I will land there and update this back to use the main branch once that
happens.