Add shared memory barriers for RAW dependencies (#90)
This PR introduces a shared memory barrier op that can be used to to
introduce barriers between writes and reads to shared memory. We use the
write_dependency field of read nodes as well as the node indices to
determine whether to insert a barrier. Also, we try to re-use barriers
to minimize the total number of barriers.

---------

Signed-off-by: Harsh Menon <harsh@nod-labs.com>