[TKW] Implement define interface op and tk.max
In this PR, we add 3 things:

1. Define Interface Op
2. Reduction max op
3. elementwise maximum op

Define interface op is created to generate new subclass op that can
derive from base interface class. this is useful for custom TK ops that
has very similar properties such as "sum" and "max" S.T we can reuse
code as much as possible.

Signed-off-by: Stanley Winata <stanley.winata@amd.com>