Add e2e parameterization using file (#119)
This PR and future related PRs will be transformed into performance-CI,
benchmarking different operators with various input parameters. For now,
we are starting with GEMMs. In the future, we will add convolution,
flash attention, and other operators as needed. The ultimate goal is to
create performance plots for those operators so we can see their
performance daily.

The perf.yaml file is added to the repo to distinguish between
Performance and CI related operations.

---------

Signed-off-by: erman-gurses <erman@nod-labs.com>