[TKW] Add register analysis support for ReduceOp (#150)
Implement pieces to enable partial reductions/tiling of
ReduceOps(max/sum) in a for loop/tkw.reduction loop.

1.  Add register analysis support for reduce op.
2. Updated implementation of `replace_all_uses_with` with FX's
implementation of it.

num.2 is required to support updating the use of result from `ReduceOp`
on `Output` from `Reduction`. In this case, with existing implementation
of `update_arg(user.get_node_arg_index(` where `user` is `Output` causes
issues because `output.fx_node.arg` is a `tuple[list[fx.node]]`, we can
update it to not use fx's implementation but may get quite invasive and
ad-hoc.

---------

Signed-off-by: Stanley Winata <stanley.winata@amd.com>
Signed-off-by: Ian <ian.nordeng@amd.com>