[TKW] Thread Shape analysis
The motivation of this pass is to generalize the register analysis pass
which is used to determine the thread shape of TKW.Register, to all
other operations.

One main use case for such is to allow reduction, and later on
"broadcast" to use thread shape information from the kernel as opposed
to relying on vector_shape which may not always be valid.

We generalize the register analysis metho by finding  a few anchor ops
who's thread shape information is determined,
and then propagate to it's successors and ancestors.

Signed-off-by: Stanley Winata <stanley.winata@amd.com>