Add in-place variant of transfer_to_logical_device
The purpose of this is to insert IREE ABI affinity attributes for
function arguments.
iree.flow.transfer is not sufficient for cases of in-place updates of
function arguments.

An alternative approach is to set the attribute of a Torch FX graph
function argument. Then during/after conversion to MLIR to set the
corresponding attribute on the MLIR function.

Signed-off-by: Boian Petkantchin <boian.petkantchin@amd.com>