[TKW] Fix types, shapes and propagate resolved indexing  (#177)
This PR add supports for doing accumulate on non-induction variables.
For most part our stack already supports this, but we'd need to fix
reduction symbolic shape which used to be just the input shape to input
shape - reduction dim. Without this our thread shape analysis wouldn't
be able to handle broadcasting of reduced op properly.

To support the above, we also introduce `Extract` op in addition to the
existing `ExtractSlice` to fix the shape types. Specifically,
ExtractSlice follow upstream's semantic where we just slice but do not
reduce any dimensions. For the ReduceOp case, specifically the local
reduction, we'd want it to have a reducing semantic on the fastest
dimension.

Additionally, we also add a propagation of our resolution for thread
shape. This is helpful in the test we added which is a broadcast-sub
followed by an exp2.

---------

Signed-off-by: Stanley Winata <stanley.winata@amd.com>