[TKW] Propagate GetResult/IterArg of Reductions out for expansion and thread shape (#225)
In this PR, we add support to propagate indexing, expansion, and thread
shape information from IterArg/induction variable of Reduction loop to
it's user outside the loop.

This is important to enable for example having two induction variables
in a reduction loop with different layout/index, and we'd want to do a
binaryOp on them before writing them out. We'd need to figure if for
example it'd need a broadcast or not. That won't be possible if we do
not propagate this information from IterArg to outside/GetResult and
it's use-def chain.

---------

Signed-off-by: Stanley Winata <stanley.winata@amd.com>