Make transfer_to_logical_device copy the tensor during eager execution
Currently eager and generate don't have the same semantics.
Generate would create a new value while in eager we return the same
tensor.
It is desirable to have the same behavior in both paths, this will allow
us to catch in-place errors on the torch level and not during IREE
execution.
The only downside is that during eager it will be slower. This is far
less of headache.

Signed-off-by: Boian Petkantchin <boian.petkantchin@amd.com>