Add support for varying vector shapes
This PR adds support for expanding operators
with varying vector shapes, specifically for the
MMA case where either the same dimension has different
vector shapes in different mmas or if different instructions
are being used.

The idea is to insert a reshape operator whenever such a shape
mismatch is discovered.  The reshape operator lowers to an extract or
concatenate operation, depending on the context.

Signed-off-by: Harsh Menon <harsh@nod-labs.com>