[TKW] Teach expansion to handle non direct acc and ReduceOp on reduction dim. (#243)
In flash attention, we need to enable non direct acc matmul, and also
expansion of reduceOp in reduction dimension. The former is needed in FA
since we are applying some scaling to the acc of second MMA before
feeding it in. The second case is required in FA because ReduceOp/MaxOp
is in the backward slice of second MMA's LHS, which would require it to
be expanded in K2/reduction dim as well.

---------

Signed-off-by: Stanley Winata <stanley.winata@amd.com>