Add type inference (#252)
This PR adds a type inference pass to wave. Previously, the types were
infered by looking up types from neighbors resulting in inefficient type
inference.

Instead, we now introduce a pass that infers the types for all operators
in the graph and the inferred type is then stores in the node. New nodes
that are constructed in downstream passes are responsible for annotating
types for the new operators.

---------

Signed-off-by: Harsh Menon <harsh@nod-labs.com>