Bump version to 2.5.0 and document release process. (#256)
The `docs/releasing.md` file was first introduced next to the
`build_release.py` script in
https://github.com/nod-ai/SHARK-ModelDev/pull/605. I don't see that file
in this repository (iree-org/iree-turbine) though, despite
https://github.com/nod-ai/SHARK-ModelDev/pull/655 claiming that it had
been upstreamed. Since the packages involved have changed over the
months (shark-turbine --> iree-turbine naming, see also
https://github.com/iree-org/iree-turbine/issues/28), I also updated the
document to reflect the current setup.

The `iree-requirements.txt` file was being used to pin versions of the
`iree-compiler` and `iree-runtime` that would be released together with
this `iree-turbine` package. That was overlooked when
https://github.com/iree-org/iree-turbine/pull/115 removed pins from that
file. I've reworked the script to reference `iree-requirements-ci.txt`
for now, but we can continue to iterate on the cross-project release
process going forward (with docs this time).

I've updated the version in code from 2.3.0 to 2.5.0. We can push a
release to https://github.com/iree-org/iree-turbine/releases and
https://pypi.org/project/iree-turbine/#history using the new version
next. Note that 2.3.1, 2.4.0, and 2.4.1 were released to PyPI without
updating the code or GitHub releases.