[TKW] Kernel Cacher
For eager mode to be viable, we'd need to implement Kernel Cacher S.T we
do not need to re-compile kernels every time. Here are the main changes:

1. Refactor wave.py's compile_and_invoke to two separate functions
   `compile_to_vmfb`, and `invoke_vmfb` this is S.T we can intercept the
compiled vmfb cleanly and store it to the caches.
2. Implement kernel cache dataclass which is a struct necessary to
   reconstruct kernels S.T it is invokable as original state
3. Implement fn to invoke kernel cache
4. Implement kernel cache manager that can hash, load/store kernel to RAM,
   load/store kernel to files.
5. Tests and helper fn for cache manager.

Let's discuss a little more about the newly developed cache manager, the
Wave  cache manager has two main components/cache:

1. Session/Online cache - This is the main cache that our compiler and runtime
will load from and store to. It is essentially a dict that uses the kernel hash
as keys and the WaveCache as values. We added LRU functionality with limits for
number of kernel cached here, because this lives on RAM, and we wouldn't want to run OOM.

2. File/Offline cache - This cache is essential for loading saved/compiled cache
between sessions/runs. This is done by storing vital kernel information
(vmfb, kernel_sig,and mlir) to CACHE_BASE_DIR/kernel_hash directory. If said kernel
is queried during a new run and does not exist on session/online cache yet, we'd load
files from the kernel_hash directory and reconstruct the WaveCache from it.

Signed-off-by: Stanley Winata <stanley.winata@amd.com>