Redo expansion (#357)
This PR is a complete overhaul of the expansion process in wave
incorporating learnings from the past several months. Below, we list
some of the major changes -

1. Use an iterative rather than recursive approach which should allow
for better debugging
2. Allow expansion to move through reductions and no special case
expansions for reductions (previously, expansion was bfs in the
reduction and dfs elsewhere)
3. Updated printing of nodes to specify dimensions
4. Split expansion into expansion and fixup phase (reductions and MMAs
are adjusted in the fixup)
5. Only special handling for reshape & reduce ops that require
concatenation of inputs
6. Ordering of nodes is now 00, 01, 10, 11, unlike previously
7. Nodes are always cloned in expansion where previously the first
wasn't while the rest were

and many more. While updating the lit tests, I noticed that we were
doing some unnecessary reductions which are not happening anymore.

---------

Signed-off-by: Harsh Menon <harsh@nod-labs.com>