Make expansion start from all write nodes
Right now, it only looks at write nodes in the root graph and misses
ones in reductions. Instead we walk all the regions for `Write` nodes
and `GetResult` nodes without users.

`get_leaf_node_types` also appeared to always return `[Write]` as it
sets `leaf_nodes` to that and then immediately checks if it's empty,
which it never will be, so I deleted it.

Signed-off-by: Geoffrey Martin-Noble <gcmn@brium.ai>