Ensure argument ordering in graph is respected
Previously `add_from_graph_placeholders` uses the topologcal ordering of
placeholder to determine ordering of arguments/bindings. This is flaky
as it breaks during decode_attention.py tests if we do some reordering.
In this commit, we add an arg_id in the meta of the node during tracing,
S.T in `add_fomr_graph_placeholders` we can match the argument
ordering with the fx.nodes.

Signed-off-by: Stanley Winata <stanley.winata@amd.com>