Mark args as outputs if *any* user is a write (#422)
Input arguments are marked as read-only. Therefore, if there's any write
the argument has to be marked as an output.

Includes a typo bug fix for an issue that reveals itself if all
arguments are outputs.

Now that https://github.com/iree-org/iree-turbine/pull/400 made the
search for users include reductions, this fixes most of the issues noted
in https://github.com/iree-org/iree-turbine/issues/375 (reordering and
removing arguments is still a problem).