Collection of QoL changes from my efforts
This is a bad PR in the sense that it does a bunch of different things.
This is a collection of changes I made, mostly around improving
debugging with more verbose and specific error messages. But there's
also some other stuff in here. Most of these don't really seem
sufficiently important to break out into their own PR, but it seemed a
shame to just throw away things that seemed like improvements. I did
try to clean things up a bit to make them actually upstreamable.

One note is that I'm not really sure what the convention is for
Exception types in the project, so a lot of these are just RuntimeError.
That's not awesome, but I think it's still a lot more helpful than
nothing.