Prefill Attention (#412)
This PR adds an implementation of prefill attention. In prefill
attention, we have variable length sequences and so we need to pass in
the start offsets as well as the lengths of the sequences. The setting
of the offsets and variable sequence lengths is accomplished using
set_symbol.

Also, this PR adds the ability to specify the total count for a work
group constraint and adds attention cleanups.

---------

Signed-off-by: Harsh Menon <harsh@nod-labs.com>