Initial API docs config using Sphinx and readthedocs. (#434)
Progress on https://github.com/iree-org/iree-turbine/issues/77. There is
still plenty to iterate on, but this gets something started that
developers can start using and improving.

> [!IMPORTANT]
> **A demo website is hosted at
https://iree-turbine-scotttoddtest.readthedocs.io/en/latest to help with
review.**

I also claimed https://app.readthedocs.org/projects/iree-turbine/ and
will configure it once this is merged. The final hosted site will be at
https://iree-turbine.readthedocs.io/en/latest/ and will update after
each commit to the `main` branch in this repository.

## Notes

* I've structured the site into a few main sections:
  * API reference material
    * "IREE Turbine Core"
    * "TK and TKW"
  * Project documentation
    * Infrastructure
* Some module setups are conflicting with how
[`autodoc`](https://www.sphinx-doc.org/en/master/usage/extensions/autodoc.html)
wants modules to be defined. I worked with what we have now, but we may
want to change how modules are defined and imports are used. I'm fairly
inexperienced at this level, but I think a reliance on relative imports
like these is leading to packages that can't easily be used in
isolation:
https://github.com/iree-org/iree-turbine/blob/a7c333236f37fefffd513e8b97eac4085d0bf72e/iree/turbine/kernel/wave/codegen.py#L83-L94

This show up in the docs where something like `automodule::
iree.turbine.aot.builtins` for a given module is missing significant
classes and functions and code paths like
https://github.com/iree-org/iree-turbine/tree/main/iree/turbine/kernel/compiler
have no `__init__.py` to create a package.