[Wave] Teach compiler to handle batched MMA for VMFMAOps (#495)
This handles support of VMFMA for broadcasted batched gemm. i.e the
batch size of one of the operand is broadcasted onto the other. i.e this
is to support virtual MMA intrinsics in GQA.

For example in a GQA like gemm we will have (B_KV, M, K) * (B, K, N) ->
(B, M, N). In this case, with set operation alone it is hard to see that
`B_KV` is also a batch dimension and not M-dim.

Initial prototype adds an additional set intersection with
acc_index.type. Whilst that would work for some cases, it wouldn't work
for the operand with the same batch size as the acc/output. i.e in our
case above we will fail to determine the right N-dim.
 m-dim = (B_KV, M, K)  - (B, K, N)  ∩  (B, M, N) = (M,) - OK
 n-dim =  (B, K, N)  - (B_KV, M, K)  ∩  (B, M, N) = (B, N) - NG
 
Hence, we added a filter that ensures the mma indexing expression can be
found in said dimension, making it more robust to false batch/m-dims in
the future.

---------

Signed-off-by: Stanley Winata <stanley.winata@amd.com>