Global to shared gathers (#496)
This PR adds support for transforming gathers from global memory to
gather to shared memory. In order to do this,
we modify the loads from global and stores to shared to use all the
threads available and then use gathers to obtain
the data from shared memory.

We add a new pass that closely follows minimize global loads as well as
other changes to enable this.

Signed-off-by: Harsh Menon <harsh@nod-labs.com>