Extend attention + RPE (#504)
This PR starts from an Extend attention + RPE by @raikonenfnu.

Then, avoid flex_attention for RPE as it is unclear there is a correct
implementation possible given the
limitations of the create_block_mask re conditionals.

Instead we use the manual torch implementation that is known to be
correct.

After debugging and further changes, the RPE implementation works as
expected.

Signed-off-by: Nicolas Vasilache
<nicolasvasilache@users.noreply.github.com>

---------

Signed-off-by: Stanley Winata <stanley.winata@amd.com>
Signed-off-by: Nicolas Vasilache <nicolasvasilache@users.noreply.github.com>
Signed-off-by: Alex Zinenko <alex@brium.ai>
Signed-off-by: Ivan Butygin <ivan.butygin@gmail.com>
Co-authored-by: Stanley Winata <stanley.winata@amd.com>
Co-authored-by: Alex Zinenko <alex@brium.ai>
Co-authored-by: Ivan Butygin <ivan.butygin@gmail.com>