Use LRU cache when retrieving kernel hash
This PR annotates the get_hash function for retrieving
cached kernels with functools.lru_cache to ensure that
we don't pay the cost for computing the hash for a
particular kernel more than once per session.

It also modifies the arguments to the get_hash function
to ensure that they are hashable.

Signed-off-by: Harsh Menon <harsh@nod-labs.com>