Use LRU cache when retrieving kernel hash (#528)
This PR annotates the get_hash function for retrieving cached kernels
with functools.lru_cache to ensure that we don't pay the cost for
computing the hash for a particular kernel more than once per session.

get_hash used to take 2.5ms, and with this PR that goes down to 0.9ms ,
since we are avoiding the hash calculation and just doing a lookup.

Signed-off-by: Harsh Menon <harsh@nod-labs.com>