Test reference implementations separately
These were complicating the tests and making them much less readable. I
thought that the loop-based implementations were adding significant
latency and we needed to optimize that (including only using them for
small shapes), but they turn out to be reasonably fast and negligible
compared to the Wave implementation latency. Using these as the
reference for the Wave kernels gives us a reference that's actually for
the same algorithm and includes all the intermediates.