Inject cache directory into the cache manager
The current setup hard codes the cache directory for the global cache
manager based on an environment variable. This is a problem particularly
in tests because it causes the tests to be non-hermetic and to mess with
the users cache directory.

I think it would be better if the tests could instantiate their own
cache manager explicity rather than relying on the global, but that
would require a more significant restructuring.