Also return asm on cached path
I renamed the property `asm` as that's actually what it is. I thought
this was already returning the asm, but it's returning a WaveCache
dataclass. Another option would be to consistently return a dataclass
that has the extra information, but I didn't want to use a cache type
for the path where the cache is turned off and `WaveCache` also fetches
the asm lazily, which complicates things a bit. I think we'd want a
`WaveKernel` type and two duck-typed implementations.