[TKW] Consistently return asm in test execution (#574)
Cached kernels were returning the asm and uncached ones were returning
the module builder.

I added cache statistics to the `WaveCacheManager` instead of the cache
test using the returned type to distinguish a cache hit. I also renamed
the `module_op` property on `WaveCache` to `asm as that's actually what
it is (treating it like it was a module op was what was causing failures
in the first place).

Another option would be to consistently return a dataclass that has the
extra information present in `WaveCache`, but I didn't want to use a
cache type the path where the cache is turned off and `WaveCache` also
fetches asm lazily, which complicates things a bit. I think we'd want a
`WaveKernel` type and two duck-typed implementations if we went that
route.

Fixes https://github.com/iree-org/iree-turbine/issues/419