Keep dimension naming convention during scheduling
This was creating new nodes that somehow ended up with names like K_1 to
indicate expansion dimensions instead of K:1. This was further confused
because I think if you already had nodes with the same name, then the fx
graph would increment the trailing number, so you'd get K_3 even if
there wasn't an expansion of that size in the K dimension.

I've changed it to explicitly pass the old node name when constructing
new graphs during scheduling. Given that we are replacing the previous
graph, I think having the names be the same is not a problem and names
in fx graphs can conflict and aren't load bearing, so it seems unlikely
this will break something.