[Wave] Add wave runtime (#587)
This PR adds a custom wave runtime that can be used to launch kernels.
The runtime uses nanobind to transfer the kernel arguments from Python
to C++ and then launches the kernels using the hip API. In order to
enable this, we add the C++ code for wave runtime and link against
nanobind and torch to generate a shared library that can be loaded.

For this runtime, we need the hsaco binary instead of the vmfb. So we
cache the hsaco and associated kernel launch info when the cache is
enabled. Finally, we add a pass to compute the shared memory usage by
summing up the bytes used by all the shared memory allocations.

This PR will require all old cache directories to be removed.

---------

Signed-off-by: Harsh Menon <harsh@nod-labs.com>