[TKW] Keep disambiguating number in expanded node name (#437)
Nodes get named like `add`, `add_1`, etc to disambiguate ones with the
same name. But the node naming in expansion throws away this suffix. I
think it's helpful to leave it in, when viewing the IR. Nodes don't have
to have unique names, but it's nice if they do.

In updating the lit tests, I also noticed that the node names in
scheduling.py had a different convention for dimension expansion than
everything else, writing it `K_1` instead of `K:1`. To further confuse
things, it looked like torch was actually incrementing that trailing
`_1` if node names conflicted. Since my change made things not conflict
it appeared to change the expansion dimension number! I changed it so
that scheduling instead preserves the original node name. Given that we
are copying these nodes into a new graph, I think it is fine to repeat
node names (and in fact it's better because then you can relate them).
Node names also shouldn't be load bearing for anything and conflicts
shouldn't cause problems, so this shouldn't have the potential to create
any bugs or failures (famous last words).