[TKW] Flash Attention 2 Backward Pass (#553)
This adds a test for the forward and backward pass of attention using
the Flash Attention 2 algorithm. It includes both a unified kernel for
the backward pass and separate kernels for each gradient. Currently,
everything works only for the 16x16x16 MFMA variant. The 32x8x8 variant
works for some parts but not others. The kernels include workarounds for
a number of bugs that I've filed.

I think this is in a state where we can check it in and continue
iterating on it.