[Wave] Add 2-stage prefetch scheduling strategy for GEMM (#614)
This PR is a MVP/first bring up for a heuristics based scheduling
strategy into Wave. In this PR, we introduce 2-stage prefetch scheduling
(strategy currently only tested for GEMM). Just for a brief recap, what
we mean by 2 stage prefetch strategy is as such:

```
    Convert vanilla schedule of:
        for i = 0 to N:
            a = READ_GLOBAL i
            WRITE_SHARED a
            barrier
            b = READ_SHARED
            COMPUTE b

    into prefetch schedule:
        a_0 = READ_GLOBAL 0
        WRITE_SHARED a_0
        for i = 0 to N - 1:
            a_{i+1} = READ_GLOBAL i + 1
            // a_{i+1} is NOT blocked by this barrier because barriers only block shared memory transfers
            barrier
            b_i = READ_SHARED
            COMPUTE b_i
            barrier
            WRITE_SHARED a_{i+1}
        barrier
        b_N = READ_SHARED
```

In order to do that we needed to bring these changes:

1. Implement a scheduler class that generate a schedule and initiation
interval that generates a 2 stage prefetch strategy. The crux of this
implementation is:
a. we'd need to sort the nodes inside `reduction_graph` but it'
accumulative scheduling weight (accumulative on it's predecessors), this
should be rather similar to topological graph but more robust since it
is based on scheduling edges weight.
b. After the nodes are sorted, we classify them to one of the 4 stages
in prefetch strategy which is `GLOBAL_LOAD`, `LOCAL_STORE`,
`LOCAL_LOAD`, and `COMPUTE`, where each stage will correspond to a clock
cycle.
c. Once this is done, we would simply set initiation interval to 2 which
should let the loop reconstructor to generate a 2 stage prefetch
strategy.
 
2. Modify scheduling/resources.py's implementation of
`get_custom_operation_type` to index on shared reads, since global reads
cannot be found in `reduction_graph`, which is the form used by prefetch
scheduler (and other schedulers for that matter).

3. General infra update and tests

Future work:
- Currently performance does not improve, but big hunch we'd just need
to remove a redundant lds barrier to unlock perf.
- Another perf factor is to add set_prio instruction for better time
slicing / resource overlap/use in a single SIMD.
- Apply to attention and extend attention

---------

Signed-off-by: Stanley Winata <stanley.winata@amd.com>